Preprocessing Stage Block Update
main.ipynb:
The new preprocessing block of code is universal for binary classification, that is, it works for any dataset, as long as its rows include only the two labels that will be used for traning.
The new algorithm uses pandas to deal with the files as it puts their data inside pandas data structure DataFrame. Also, scikit-learn's StandardScale class from its module preprocessing performs a standardization of the features data to help the traning process. Finally, scikit-learn's train_test_split class splits the data into X_train, X_test, y_train and y_test. This is important for traning and validation and they are also the output of the Preprocessing stage.
At last, it is IMPORTANT to say that docstrings must be added to the functions included in the code cell and, if necessary, new details about the preprocessing algorithm must be added to its markdown cell.

main-version1.ipynb:
As I am using Git and GitHub, I do not believe it is necessary to create a backup file with the old code in case the new code fails, however, as I don't know how to do that yet, I created it  in case my hypothesis is wrong or I do any mistake that does not allow me to come back with the original code. When I learn how to do that and guarantee my modifications can be undone, I'll exclude main-version1.ipynb.

requirements.txt:
the libraries I have installed to run the code I am using now have been added to this file in order to install them in case I lose my current pip list. At the moment the only and last libraries that have been added were:
-pandas==2.2.0
-scikit-learn==1.4.1.post1

iris\iris.data:
As put before, the database in .csv format must contain rows that are only with the labels used in the classification. Therefore, I had to exclude all observations that were about Iris-virginica.